move authentication routes to middleware, add meaningful errors if no env file
